Crate merino_suggest

Suggestion backends for Merino.

Modules

debug
A suggestion provider that provides debug responses.

multi
Provides a provider-combinator that provides suggestions from multiple sub-providers.

wikifruit
A suggestion provider that provides toy responses.

Structs

DebugProvider
A toy suggester to test the system.

Language
A representation of a language, as given in the Accept-Language HTTP header.

Multi
A provider that aggregates suggestions from multiple suggesters.

Suggestion
A suggestion to provide to a user.

SuggestionRequest
A request for suggestions.

SuggestionResponse
A response of suggestions, along with related metadata.

SupportedLanguages
Languages supported by the client.

WikiFruit
A toy suggester to test the system.

Enums

CacheStatus
The relation between an object and a cache.

LanguageIdentifier
An enum used to signify whether a Language refers to a specific language or a wildcard.

SetupError
Errors that may occur while setting up the provider.

SuggestError
Errors that may occur while querying for suggestions.

Traits

SuggestionProvider
A backend that can provide suggestions for queries.

Functions

fake_example_url
Helper to generate a URL to use for testing, of the form “https://example.com/fake#some-random-words”.