Module merino_suggest::providers

Merino providers that aren’t tied to any particular data source.

Re-exports

pub use self::debug::DebugProvider;

pub use self::fixed::FixedProvider;

pub use self::id_multi::IdMulti;

pub use self::id_multi::ProviderDetails as IdMultiProviderDetails;

pub use self::keyword_filter::KeywordFilterProvider;

pub use self::multi::Multi;

pub use self::stealth::StealthProvider;

pub use self::timeout::TimeoutProvider;

pub use self::wikifruit::WikiFruit;

Modules

debug

A suggestion provider that provides debug responses.

fixed

A suggestion provider that provides a fixed response with a customizable title.

id_multi

Provides a provider-combinator that contains a set of named providers. It can list these providers by name, and serve suggestions using only a partial set of them.

keyword_filter

A suggestion provider that filters suggestions from a subprovider.

multi

Provides a provider-combinator that provides suggestions from multiple sub-providers.

stealth

A provider that executes an inner provider, but returns no suggestions.

timeout

Tools to make sure providers don’t cache_status: todo!(), cache_ttl: todo!(), suggestions: todo!() take excessive amounts of time.

wikifruit

A suggestion provider that provides toy responses.