Module merino_suggest::providers::id_multi

Provides a provider-combinator that contains a set of named providers. It can list these providers by name, and serve suggestions using only a partial set of them.

This is intended to be used only as the top level provider for the service.

Structs

IdMulti

A provider that aggregates suggestions from suggesters that tracks an ID per suggester (or suggester tree).

ProviderDetails

Metadata about a provider contained in [NamedMulti];

Enums

ProviderAvailability

The availability of a provider, which affects if it is used by default in requests, and how clients can configure it.