Trait merino_suggest::SuggestionProvider

pub trait SuggestionProvider: Send + Sync {
    fn name(&self) -> String;

    fn suggest<'life0, 'async_trait>(
        &'life0 self, 
        query: SuggestionRequest
    ) -> Pin<Box<dyn Future<Output = Result<SuggestionResponse, SuggestError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn is_null(&self) -> bool { ... }
}

A backend that can provide suggestions for queries.

Required methods

fn name(&self) -> String

An operator-visible name for this suggestion provider.

fn suggest<'life0, 'async_trait>(
&'life0 self,
query: SuggestionRequest
) -> Pin<Box<dyn Future<Output = Result<SuggestionResponse, SuggestError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Provide suggested results for query.

Provided methods

fn is_null(&self) -> bool

Return if this provider is null and can be ignored. Providers that set this to true should be ignored in any place where suggestions are needed. Providers with this set to true likely only serve as a blank space where we may need a provider but can’t otherwise supply one.